Responsive The home Page

+++ src/Components/Pages/Dashboard/ManageOrders.js
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from "react";
 import { toast } from "react-toastify";
 import "./Dashboard.css";
+import { confirmAlert } from "react-confirm-alert";
 
 const ManageOrders = () => {
     const [orders, setOrders] = useState([]);
     useEffect(() => {
-        fetch("http://localhost:5000/orders/admin", {
+        fetch("https://dry-forest-04223.herokuapp.com/orders/admin", {
             method: "GET",
             headers: {
                 authorization: `Bearer ${localStorage.getItem("accessToken")}`,
@@ -15,7 +16,7 @@ const ManageOrders = () => {
             .then((data) => setOrders(data));
     }, []);
     const handleShipment = (id) => {
-        fetch(`http://localhost:5000/order/${id}`, {
+        fetch(`https://dry-forest-04223.herokuapp.com/order/${id}`, {
             method: "PATCH",
             headers: {
                 "content-type": "application/json",
@@ -31,18 +32,40 @@ const ManageOrders = () => {
             });
     };
     const handleCancel = (id) => {
-        fetch(`http://localhost:5000/order/${id}`, {
-            method: "DELETE",
-            headers: {
-                authorization: `Bearer ${localStorage.getItem("accessToken")}`,
-            },
-        })
-            .then((res) => res.json())
-            .then((data) => {
-                if (data.acknowledged) {
-                    toast("Order Delete Successfully");
-                }
-            });
+        confirmAlert({
+            title: "Confirm to submit",
+            message: "Are you sure to do this.",
+            buttons: [
+                {
+                    label: "Yes",
+                    onClick: () => {
+                        fetch(
+                            `https://dry-forest-04223.herokuapp.com/order/${id}`,
+                            {
+                                method: "DELETE",
+                                headers: {
+                                    authorization: `Bearer ${localStorage.getItem(
+                                        "accessToken"
+                                    )}`,
+                                },
+                            }
+                        )
+                            .then((res) => res.json())
+                            .then((data) => {
+                                if (data.acknowledged) {
+                                    toast("Order Delete Successfully");
+                                }
+                            });
+                    },
+                },
+                {
+                    label: "No",
+                    onClick: () => {
+                        return;
+                    },
+                },
+            ],
+        });
     };
     return (
         <div className="m-2">
